Modificado Layout de RegistroActivity, add lambda, add texto jundo ao bitmap

+++ app/src/main/java/com/example/controledeextintor/DbDados.java
@@ -2,14 +2,18 @@ package com.example.controledeextintor;
 
 import android.content.ContentValues;
 import android.content.Context;
+import android.content.Intent;
 import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.icu.util.Calendar;
+import android.net.Uri;
 import android.provider.BaseColumns;
 import android.widget.Toast;
 
+import java.io.File;
+import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
@@ -31,10 +35,11 @@ public class DbDados extends SQLiteOpenHelper {
                 + QrCodeEntry._ID + " INTEGER PRIMARY KEY AUTOINCREMENT, "
                 + QrCodeEntry.COLUMN_QRCODE_CLASS + " TEXT NOT NULL, "
                 + QrCodeEntry.COLUMN_QRCODE_SETOR + " TEXT NOT NULL, "
+                + QrCodeEntry.COLUMN_QRCODE_TIPO + " TEXT NOT NULL, "
+                + QrCodeEntry.COLUMN_QRCODE_MEDIDA + " TEXT NOT NULL, "
                 + QrCodeEntry.COLUMN_QRCODE_DATE_RECARGA + " TEXT NOT NULL, "
                 + QrCodeEntry.COLUMN_QRCODE_DATE_INSPECAO + " TEXT NOT NULL,"
                 + QrCodeEntry.COLUMN_QRCODE_DATE_QRCODE + " TEXT NOT NULL);";
-        ;
 
         db.execSQL(SQL_CREATE_QRCODE_TABLE);
     }
@@ -49,28 +54,23 @@ public class DbDados extends SQLiteOpenHelper {
         // Implemente este método se quiser rebaixar o banco de dados
     }
 
-    public void insertData(String qrCodeClass, String qrCodeSetor) {
+    public void insertData(String qrCodeClass, String qrCodeSetor, String qrCodeTipo, String qrCodeMedida, String qrCodeDateRecarga, String qrCodeDateInspecao) {
         SQLiteDatabase db = this.getWritableDatabase();
         ContentValues contentValues = new ContentValues();
 
         contentValues.put(QrCodeEntry.COLUMN_QRCODE_CLASS, qrCodeClass);
         contentValues.put(QrCodeEntry.COLUMN_QRCODE_SETOR, qrCodeSetor);
+        contentValues.put(QrCodeEntry.COLUMN_QRCODE_TIPO, qrCodeTipo);
+        contentValues.put(QrCodeEntry.COLUMN_QRCODE_MEDIDA, qrCodeMedida);
 
         // Obtenha a data atual em formato de string com o padrão desejado
         SimpleDateFormat dateFormat = new SimpleDateFormat("dd-MM-yyyy");
-        String currentDate = dateFormat.format(new Date());
 
         // Obtenha a data corrente
         Calendar calendar = Calendar.getInstance();
         calendar.setTime(new Date());
 
-        // Adicione 1 ano à data corrente para a data de recarga
-        calendar.add(Calendar.YEAR, 1);
-        String qrCodeDateRecarga = dateFormat.format(calendar.getTime());
-
-        // Adicione 5 anos à data corrente para a data de inspeção
-        calendar.add(Calendar.YEAR, 5);
-        String qrCodeDateInspecao = dateFormat.format(calendar.getTime());
+        String currentDate = dateFormat.format(new Date());
 
         // Insira a data de troca
         contentValues.put(QrCodeEntry.COLUMN_QRCODE_DATE_RECARGA, qrCodeDateRecarga);
@@ -84,6 +84,8 @@ public class DbDados extends SQLiteOpenHelper {
         db.insert(QrCodeEntry.TABLE_NAME, null, contentValues);
         db.close();
     }
+
+
     public String buscarDados() {
         StringBuffer dados = new StringBuffer();
 
@@ -102,12 +104,14 @@ public class DbDados extends SQLiteOpenHelper {
                     int id = cursor.getInt(cursor.getColumnIndex(QrCodeEntry._ID));
                     String qrCodeClass = cursor.getString(cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_CLASS));
                     String qrCodeSetor = cursor.getString(cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_SETOR));
+                    String qrCodeTipo = cursor.getString(cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_TIPO));
+                    String qrCodeMedida = cursor.getString(cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_MEDIDA));
                     String qrCodeDateRecarga = cursor.getString(cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_RECARGA));
                     String qrCodeDateInspecao = cursor.getString(cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_INSPECAO));
                     String qrCodeDateQrcode = cursor.getString(cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_QRCODE));
 
                     // Formata os dados no formato desejado
-                    String dadosSalvos = "mp_" + id + ";" + qrCodeClass + ";" + qrCodeSetor + ";" + qrCodeDateRecarga + ";" + qrCodeDateInspecao + ";" + qrCodeDateQrcode + "\n";
+                    String dadosSalvos = "mp_" + id + ";" + qrCodeClass + ";" + qrCodeSetor + ";" + qrCodeTipo + ";" + qrCodeMedida+ ";" + qrCodeDateRecarga + ";" + qrCodeDateInspecao + ";" + qrCodeDateQrcode + "\n";
 
                     // Adiciona os dados formatados ao StringBuffer
                     dados.append(dadosSalvos);
@@ -121,6 +125,7 @@ public class DbDados extends SQLiteOpenHelper {
 
         return dados.toString();
     }
+
     public List<String> getDistinctProductNames(String searchText) {
         // Obtém uma instância legível do banco de dados
         SQLiteDatabase db = this.getReadableDatabase();
@@ -129,7 +134,7 @@ public class DbDados extends SQLiteOpenHelper {
         List<String> productList = new ArrayList<>();
 
         // Define as colunas que serão retornadas na consulta
-        String[] columns = new String[]{QrCodeEntry._ID, QrCodeEntry.COLUMN_QRCODE_SETOR, QrCodeEntry.COLUMN_QRCODE_CLASS, QrCodeEntry.COLUMN_QRCODE_DATE_RECARGA, QrCodeEntry.COLUMN_QRCODE_DATE_INSPECAO, QrCodeEntry.COLUMN_QRCODE_DATE_QRCODE};
+        String[] columns = new String[]{QrCodeEntry._ID, QrCodeEntry.COLUMN_QRCODE_SETOR, QrCodeEntry.COLUMN_QRCODE_CLASS, QrCodeEntry.COLUMN_QRCODE_TIPO, QrCodeEntry.COLUMN_QRCODE_MEDIDA, QrCodeEntry.COLUMN_QRCODE_DATE_RECARGA, QrCodeEntry.COLUMN_QRCODE_DATE_INSPECAO, QrCodeEntry.COLUMN_QRCODE_DATE_QRCODE};
 
         // Define a cláusula WHERE para filtrar os resultados pelo termo de busca fornecido
         String selection = QrCodeEntry.COLUMN_QRCODE_SETOR + " LIKE ?";
@@ -145,27 +150,34 @@ public class DbDados extends SQLiteOpenHelper {
             int indexId = cursor.getColumnIndex(QrCodeEntry._ID);
             int indexSetor = cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_SETOR);
             int indexClass = cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_CLASS);
+            int indexTipo = cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_TIPO);
+            int indexMedida = cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_MEDIDA);
             int indexDateRecarga = cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_RECARGA);
             int indexDateInspecao = cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_INSPECAO);
             int indexDateQrcode = cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_QRCODE);
-            if (indexId >= 0 && indexSetor >= 0 && indexClass >= 0 && indexDateRecarga >= 0 && indexDateInspecao >= 0 && indexDateQrcode >= 0) {
-                do {
-                    int id = cursor.getInt(indexId);
-                    String setor = cursor.getString(indexSetor);
-                    String qrCodeClass = cursor.getString(indexClass);
-                    String qrCodeDateRecarga = cursor.getString(indexDateRecarga);
-                    String qrCodeDateInspecao = cursor.getString(indexDateInspecao);
-                    String qrCodeDateQrcode = cursor.getString(indexDateQrcode);
-                    // Concatena as colunas retornadas em uma única string
-                    String productInfo = "ID: " + id + "\n" +
-                            "Classe: " + qrCodeClass + "\n" +
-                            "Setor: " + setor + "\n" +
-                            "Data de Recarga: " + qrCodeDateRecarga + "\n" +
-                            "Data de Inspeção: " + qrCodeDateInspecao + "\n" +
-                            "Data do QRCode: " + qrCodeDateQrcode;
-                    productList.add(productInfo);
-                } while (cursor.moveToNext());
-            }
+
+            do {
+                int id = cursor.getInt(indexId);
+                String setor = cursor.getString(indexSetor);
+                String tipo = cursor.getString(indexTipo);
+                String medida = cursor.getString(indexMedida);
+                String qrCodeClass = cursor.getString(indexClass);
+                String qrCodeDateRecarga = cursor.getString(indexDateRecarga);
+                String qrCodeDateInspecao = cursor.getString(indexDateInspecao);
+                String qrCodeDateQrcode = cursor.getString(indexDateQrcode);
+
+                // Concatena as colunas retornadas em uma única string
+                String productInfo = "ID: " + id + "\n" +
+                        "Classe: " + qrCodeClass + "\n" +
+                        "Setor: " + setor + "\n" +
+                        "Tipo: " + tipo + "\n" +
+                        "Medida: " + medida + "\n" +
+                        "Data de Recarga: " + qrCodeDateRecarga + "\n" +
+                        "Data de Inspeção: " + qrCodeDateInspecao + "\n" +
+                        "Data do QRCode: " + qrCodeDateQrcode;
+
+                productList.add(productInfo);
+            } while (cursor.moveToNext());
         }
 
         // Fecha o cursor e a conexão com o banco de dados
@@ -176,10 +188,8 @@ public class DbDados extends SQLiteOpenHelper {
         return productList;
     }
 
-
-
     public String buscarDadosPorId(int id) {
-        StringBuffer dados = new StringBuffer();
+        StringBuilder dados = new StringBuilder();
 
         // Consulta para buscar os dados do banco de dados pelo ID
         String query = "SELECT " + QrCodeEntry.COLUMN_QRCODE_DATE_RECARGA + ", " + QrCodeEntry.COLUMN_QRCODE_DATE_INSPECAO +
@@ -189,15 +199,23 @@ public class DbDados extends SQLiteOpenHelper {
         SQLiteDatabase db = this.getReadableDatabase();
         Cursor cursor = db.rawQuery(query, new String[]{String.valueOf(id)});
 
-        // Verifica se o cursor é nulo
+        // Verifica se o cursor é nulo e se possui registros
         if (cursor != null && cursor.moveToFirst()) {
-            // Obtém os valores das colunas desejadas
-            String DateRecarga = cursor.getString(cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_RECARGA));
-            String DateInspecao = cursor.getString(cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_INSPECAO));
+            // Obtém os índices das colunas
+            int dateRecargaIndex = cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_RECARGA);
+            int dateInspecaoIndex = cursor.getColumnIndex(QrCodeEntry.COLUMN_QRCODE_DATE_INSPECAO);
 
-            String datas = DateRecarga + ";" + DateInspecao;
+            // Verifica se os índices são válidos (>= 0) antes de acessar os valores das colunas
+            if (dateRecargaIndex >= 0 && dateInspecaoIndex >= 0) {
+                // Obtém os valores das colunas desejadas
+                String dateRecarga = cursor.getString(dateRecargaIndex);
+                String dateInspecao = cursor.getString(dateInspecaoIndex);
 
-            dados.append(datas);
+                // Formata as datas no formato desejado
+                String datas = dateRecarga + ";" + dateInspecao;
+
+                dados.append(datas);
+            }
 
             cursor.close();
         }
@@ -206,6 +224,7 @@ public class DbDados extends SQLiteOpenHelper {
 
         return dados.toString();
     }
+
 
     public void deletarBancoDados() {
         SQLiteDatabase db = this.getWritableDatabase();
@@ -219,17 +238,19 @@ public class DbDados extends SQLiteOpenHelper {
         db.close();
     }
 
-    public static void delete(Context context, DbDados dbHelper, String value) {
-        // Recebe como parâmetros o contexto da aplicação, o helper do banco de dados, o valor e a data do registro a ser deletado
-        SQLiteDatabase db = null;
+    public static void delete(DbDados dbHelper, Context context, String value) {
+        // Este código realiza uma consulta ao banco de dados SQLite para obter o
+        // registro com um determinado ID (value) da tabela QrCodeEntry.
+        SQLiteDatabase db = dbHelper.getWritableDatabase();
         try {
-            // Obtém uma instância do banco de dados em modo escrita
-            db = dbHelper.getWritableDatabase();
-            // Define a cláusula WHERE para selecionar o registro a ser deletado
-            String whereClause = DbDados.QrCodeEntry.COLUMN_QRCODE_SETOR + " = ?";
-            // Define os argumentos da cláusula WHERE
+            // É criada uma cláusula de consulta (whereClause) que especifica que o
+            // ID deve ser igual ao valor fornecido.
+            String whereClause = DbDados.QrCodeEntry._ID + " = ?";
+            // Os valores da cláusula de consulta são fornecidos em um array de strings (whereArgs).
             String[] whereArgs = {value};
-            // Executa a query para buscar o ID do primeiro registro encontrado com base nos argumentos informados
+            // A consulta retorna um cursor que contém o registro correspondente. Os campos retornados são especificados
+            // no segundo parâmetro do método "query".
+            // O cursor é retornado para que possa ser usado para recuperar os valores dos campos do registro.
             Cursor cursor = db.query(DbDados.QrCodeEntry.TABLE_NAME,
                     new String[]{DbDados.QrCodeEntry._ID},
                     whereClause,
@@ -238,26 +259,21 @@ public class DbDados extends SQLiteOpenHelper {
                     null,
                     null,
                     "1");
-            // Verifica se o cursor retornou algum registro
-            if (cursor.moveToFirst()) {
-                // Obtém o ID do registro retornado pelo cursor
-                long id = cursor.getLong(cursor.getColumnIndex(DbDados.QrCodeEntry._ID));
-                // Deleta o registro do banco de dados com base no ID obtido
+            if (cursor != null && cursor.moveToFirst()) {
+                long id = cursor.getLong(cursor.getColumnIndexOrThrow(DbDados.QrCodeEntry._ID));
                 int rowsDeleted = db.delete(DbDados.QrCodeEntry.TABLE_NAME, DbDados.QrCodeEntry._ID + " = ?", new String[]{String.valueOf(id)});
-                // Verifica se o registro foi deletado com sucesso
                 if (rowsDeleted > 0) {
                     Toast.makeText(context, "Registro deletado com sucesso", Toast.LENGTH_SHORT).show();
                 } else {
-                    Toast.makeText(context,"Nenhum registro encontrado para deletar", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(context, "Nenhum registro encontrado para deletar", Toast.LENGTH_SHORT).show();
                 }
             } else {
-                Toast.makeText(context,"Nenhum registro encontrado para deletar", Toast.LENGTH_SHORT).show();
+                Toast.makeText(context, "Nenhum registro encontrado para deletar", Toast.LENGTH_SHORT).show();
             }
         } catch (SQLException e) {
             e.printStackTrace();
-            Toast.makeText(context,"Erro ao deletar registro", Toast.LENGTH_SHORT).show();
+            Toast.makeText(context, "Erro ao deletar registro", Toast.LENGTH_SHORT).show();
         } finally {
-            // Fecha a conexão com o banco de dados
             if (db != null) {
                 db.close();
             }
@@ -269,6 +285,8 @@ public class DbDados extends SQLiteOpenHelper {
         public static final String TABLE_NAME = "extintores_table";
         public static final String COLUMN_QRCODE_CLASS = "classe";
         public static final String COLUMN_QRCODE_SETOR = "local";
+        public static final String COLUMN_QRCODE_TIPO = "tipo";
+        public static final String COLUMN_QRCODE_MEDIDA = "medida";
         public static final String COLUMN_QRCODE_DATE_RECARGA = "data_recarga";
         public static final String COLUMN_QRCODE_DATE_INSPECAO = "data_inspecao";
         public static final String COLUMN_QRCODE_DATE_QRCODE = "data_qrcode";
